Changed Map Layout + Controller Menu Support
I updated the Movement test map to better reflect what we had in mind for the later game.

I also implemented controller support for the menu in the movement test, which works fully now.

I updated the look of the main menu (which inlcudes updating the interactions with the buttons and fonts), but have yet to implement controller support to the main menu.