Bugfix(s): fixed pause menu bugs, and reworked the main menu
I did a little more work on the controller / mouse bug to make it feel a little better.
I fixed the bugs that allowed the player to look around while paused, and also solved the bug where the player would jump when hitting resume.
Reworked the main menu system to solve the controller / mouse bug.