Rerouted Player Controls + Created Player Input Manager
Rerouted all player controls into the input manager in order to create a player Prefab.

Created the PlayerManager, which allows for up to 4 player prefabs to entire the scene and move around. Still working on having separated controls for each instance (every instance moves as one)

Big step to multiplayer!