Multiplayer Join Menu + Player Loading Up!!!!
There is a lot of added files here so i'm just gonna give a rundown. I added in a player join screen that allows for players to assign themselves to a controller, while also giving a visual confirmation that a player was added to the game. In the game scene, I added several files that keep track of which controller goes to which player, and where to spawn each player.

I know that it works with 2 players, and will begin to test with 3-4.